Roles & Permission (Finished)

+++ app/Http/Controllers/Api/PostController.php
@@ -2,11 +2,11 @@
 
 namespace App\Http\Controllers\Api;
 
-use App\Http\Controllers\Controller;
-use App\Http\Requests\StorePostRequest;
-use App\Http\Resources\PostResource;
 use App\Models\Post;
-use Illuminate\Http\Request;
+use App\Http\Controllers\Controller;
+use App\Http\Resources\PostResource;
+use Illuminate\Support\Facades\Gate;
+use App\Http\Requests\StorePostRequest;
 use Illuminate\Database\Eloquent\Builder;
 
 class PostController extends Controller
@@ -23,33 +23,36 @@ class PostController extends Controller
         }
 
         $posts = Post::with('category')
-        ->when(request('search_category'), function (Builder $query) {
-            $query->where('category_id', request('search_category'));
-        })
-        ->when(request('search_id'), function (Builder $query) {
-            $query->where('id', request('search_id'));
-        })
-        ->when(request('search_title'), function (Builder $query) {
-            $query->where('title', 'like', '%' . request('search_title') . '%');
-        })
-        ->when(request('search_content'), function (Builder $query) {
-            $query->where('content', 'like', '%' . request('search_content') . '%');
-        })
-        ->when(request('search_global'), function (Builder $query) {
-            $query->where(function(Builder $q) {
-                $q->where('id', request('search_global'))
-                    ->orWhere('title', 'like', '%' . request('search_global') . '%')
-                    ->orWhere('content', 'like', '%' . request('search_global') . '%');
+            ->when(request('search_category'), function (Builder $query) {
+                $query->where('category_id', request('search_category'));
+            })
+            ->when(request('search_id'), function (Builder $query) {
+                $query->where('id', request('search_id'));
+            })
+            ->when(request('search_title'), function (Builder $query) {
+                $query->where('title', 'like', '%' . request('search_title') . '%');
+            })
+            ->when(request('search_content'), function (Builder $query) {
+                $query->where('content', 'like', '%' . request('search_content') . '%');
+            })
+            ->when(request('search_global'), function (Builder $query) {
+                $query->where(function(Builder $q) {
+                    $q->where('id', request('search_global'))
+                        ->orWhere('title', 'like', '%' . request('search_global') . '%')
+                        ->orWhere('content', 'like', '%' . request('search_global') . '%');
 
-            });
-        })
-        ->orderBy($orderColumn, $orderDirection)
-        ->paginate(10);
+                });
+            })
+            ->orderBy($orderColumn, $orderDirection)
+            ->paginate(10);
+
         return PostResource::collection($posts);
     }
 
     public function store(StorePostRequest $request)
     {
+        $this->authorize('posts.create');
+
         if ($request->hasFile('thumbnail')) {
             $filename = $request->file('thumbnail')->getClientOriginalName();
             info($filename);
@@ -61,6 +64,8 @@ class PostController extends Controller
 
     public function show(Post $post)
     {
+        $this->authorize('posts.update');
+
         return new PostResource($post);
     }
 
@@ -73,6 +78,8 @@ class PostController extends Controller
 
     public function destroy(Post $post)
     {
+        $this->authorize('posts.delete');
+
         $post->delete();
 
         return response()->noContent();
